use message ID 0 for log messages
the client sends the request server info message with ID 1. if the
log messages also have ID 1, the client may interpret log messages
sent by the server during initial connect (for example, the message
which logs the connection) as reply to request server info, which
makes it break in ButtplugClient.java:169 (requestServerInfo()).